requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.17.1
dotenv==0.9.9
gunicorn==25.3.0
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
